refactor: create TableOfContentsLink
